+++ src/main/java/com/casaglass/casaglass_backend/repository/CorteRepository.java
@@ -20,16 +20,16 @@ public interface CorteRepository extends JpaRepository<Corte, Long> {
     );
 
     @Query("SELECT c FROM Corte c WHERE c.largoCm BETWEEN :largoMin AND :largoMax")
-    List<Corte> findByLargoRange(@Param("largoMin") BigDecimal largoMin, 
-                                 @Param("largoMax") BigDecimal largoMax);
+    List<Corte> findByLargoRange(@Param("largoMin") Double largoMin, 
+                                 @Param("largoMax") Double largoMax);
 
     @Query("SELECT c FROM Corte c WHERE c.precio BETWEEN :precioMin AND :precioMax")
-    List<Corte> findByPrecioRange(@Param("precioMin") BigDecimal precioMin, 
-                                  @Param("precioMax") BigDecimal precioMax);
+    List<Corte> findByPrecioRange(@Param("precioMin") Double precioMin, 
+                                  @Param("precioMax") Double precioMax);
 
-    List<Corte> findByLargoCmGreaterThanEqual(BigDecimal largoMinimo);
+    List<Corte> findByLargoCmGreaterThanEqual(Double largoMinimo);
 
-    List<Corte> findByPrecioLessThanEqual(BigDecimal precioMaximo);
+    List<Corte> findByPrecioLessThanEqual(Double precioMaximo);
 
     @Query("SELECT c FROM Corte c WHERE c.observacion IS NOT NULL AND c.observacion != ''")
     List<Corte> findCortesWithObservaciones();
